+++ BE/src/main/java/codesquad/team17/gnb/place/repository/PlaceRepository.java
@@ -1,5 +1,6 @@
-package codesquad.team17.gnb.place.domain;
+package codesquad.team17.gnb.place.repository;
 
+import codesquad.team17.gnb.place.domain.Place;
 import codesquad.team17.gnb.place.dto.PlaceQueries;
 
 import java.time.LocalDate;
